ScrollboxManager: Refactor 2
